Moved QuestionCard contents to separate files
